+++ build.rs
@@ -1,0 +1,8 @@
+extern crate winres;
+
+fn main() {
+    if cfg!(target_os = "windows") {
+        let res = winres::WindowsResource::new();
+        res.compile().unwrap();
+    }
+}